Initial commit; working with free API
This contains the very first iteration of the library. It has no dependencies and works with the "current weather" and "5 day forecast" APIs available free from OpenWeatherMap. Further testing and polishing is to come, but this is usable as far as I can tell. It does NOT have any real error checking.